backend/requirements.txt
fastapi
uvicorn
langchain
langchain-groq
python-dotenv
requests
